Modify Help command
The link for the "help" command was changed from the project's
repository to the user guide.

The "help" command now opens the user guide in the device's default
browser to enhance user experience. They now no longer need
to copy the link and paste it in the browser manually.

In the event that something goes wrong, i.e. an IOException is
thrown, the user guide's link will be displayed on the app like the
original "help" command.